Do not use a login shell for LUSER
Most books have something like "exec /bin/bash" in .bash_profile for the
building user. This results in stopping and waiting for commands when
running a a login shell. The current jhalfs way of removing .bash_profile is
suboptimal, since a user may want to login as $(LUSER) and expect that the
environment is set (without having to source .bashrc). So remove the
"-i" flag in "sudo -u $(LUSER) ..."


git-svn-id: svn://svn.linuxfromscratch.org/ALFS/jhalfs/trunk@4136 7d5e3409-5ede-0310-83dd-dbcb3a32f907