Remove /tools and /cross-tools when cleaning
If we stop a build before the "do-housekeeping" target is executed, there
may be symlinks of the type /tools->/mnt/lfs/tools. If we want to restart,
for example after fixing something else, and we run the "creatingtoolsdir"
target (or similar one in CLFS) again, it will fail. It's better to clean
those symlinks when cleaning the build directory.


git-svn-id: svn://svn.linuxfromscratch.org/ALFS/jhalfs/trunk@4137 7d5e3409-5ede-0310-83dd-dbcb3a32f907