Do not rely on /tools for setting ROOT (again)
Using the chapter ancestor is not enough. It's better to use the sect1 id
and test that it starts with ch-tools: package-manager does not have this
ancestor. Also the virtual kernel filesystems are done before chroot,
so directly test this one.

This works for the current lfs book. We'll need something new if we
change the order of pages.


git-svn-id: svn://svn.linuxfromscratch.org/ALFS/jhalfs/trunk@4140 7d5e3409-5ede-0310-83dd-dbcb3a32f907