Set the correct #! executable in creat* scriptlets
Those scriptlets need to use a temporary bash, which is in /tools/bin
for current LFS. But if we use a new method (cross-compiling + sysroot)
it ends into /bin. So we have to set the correct `#!' header
in both cases.
We rely on the pesence of a "creatingtoolsdir" sect1 to know that
bash is in /tools


git-svn-id: svn://svn.linuxfromscratch.org/ALFS/jhalfs/trunk@4143 7d5e3409-5ede-0310-83dd-dbcb3a32f907